// src/infrastructure/interfaces/movie-db.responses.ts
// Movies

export interface MoviesDBResponse {
    page: number;
    dates: Dates;
    results: Result[];
    total_pages: number;
    total_results: number;
  }
  
  export interface Result {
    adult: boolean;
    backdrop_path: null | string;
    genre_ids: number[];
    id: number;
    origin_country: string[];
    original_language: OriginalLanguage;
    original_title: string;
    title: string;
    original_name: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: Date;
    first_air_date: Date;
    name: string;
    vote_average: number;
    vote_count: number;
    video: boolean;
  }
  
  export interface Dates {
    maximum: Date;
    minimum: Date;
  }
  
  export enum OriginalLanguage {
    En = 'en',
    Es = 'es',
    Ja = 'ja',
    Ko = 'ko',
  }
  

// Movie

export interface MovieReponse {
    adult:                 boolean;
    backdrop_path:         string;
    belongs_to_collection: BelongsToCollection;
    budget:                number;
    genres:                Genre[];
    homepage:              string;
    id:                    number;
    imdb_id:               string;
    origin_country:        string[];
    original_language:     string;
    original_title:        string;
    overview:              string;
    popularity:            number;
    poster_path:           string;
    production_companies:  ProductionCompany[];
    production_countries:  ProductionCountry[];
    release_date:          Date;
    revenue:               number;
    runtime:               number;
    spoken_languages:      SpokenLanguage[];
    status:                string;
    tagline:               string;
    title:                 string;
    video:                 boolean;
    vote_average:          number;
    vote_count:            number;
}

export interface BelongsToCollection {
    id:            number;
    name:          string;
    poster_path:   string;
    backdrop_path: string;
}

export interface Genre {
    id:   number;
    name: string;
}

export interface ProductionCompany {
    id:             number;
    logo_path:      string;
    name:           string;
    origin_country: string;
}

export interface ProductionCountry {
    iso_3166_1: string;
    name:       string;
}

export interface SpokenLanguage {
    english_name: string;
    iso_639_1:    string;
    name:         string;
}



// Movie Cast

export interface MovieDBCastResponse {
    id:   number;
    cast: MovieDBCast[];
    crew: MovieDBCast[];
}

export interface MovieDBCast {
    adult:                boolean;
    gender:               number;
    id:                   number;
    known_for_department: Department;
    name:                 string;
    original_name:        string;
    popularity:           number;
    profile_path:         null | string;
    cast_id?:             number;
    character?:           string;
    credit_id:            string;
    order?:               number;
    department?:          Department;
    job?:                 string;
}

export enum Department {
    Acting = "Acting",
    Directing = "Directing",
    Writing = "Writing",
}